Cleaned up utilities
