+++ src/components/roadmap/RoadmapChart.vue
@@ -30,8 +30,7 @@
 
 import { computed, onBeforeUnmount, onMounted, ref, useTemplateRef } from 'vue'
 import { CHART_BAR_PADDING, CHART_BORDER_COLOR_PRIMARY, LIST_WIDTH, ROW_HEIGHT } from './constants'
-import type { RoadmapScale } from './types'
-import { computeDateRange, computeDaysInRange, computeStartsInRange, extendWeekStarts, xForDate } from './utils'
+import { computeDateRange, computeDaysInRange, computeStartsInRange, extendWeekStarts, xForDate, type RoadmapScale } from './roadmap-utils'
 import { useItemsStore } from '@/stores/items'
 
 const props = defineProps<{
